requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
mariadb==1.1.14
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
